refactoring searchBar error fixed
