+++ Shared/MovieListPage/Tabs/AllMoviesView.swift
@@ -8,28 +8,24 @@
 import SwiftUI
 
 struct AllMoviesView: View {
-    // 9 - Buradan gitmeli
-    //@ObservedObject var moviesViewModel: MainBackgroundViewModel
     @ObservedObject var allMoviewViewModel = AllMoviesViewModel()
-
+    @Binding var searchResults: [Movie]
+    
     var body: some View {
         ScrollView(showsIndicators: false){
             LazyVStack{
-                //ForEach(moviesViewModel.searchResults.isEmpty ?  moviesViewModel.allMovies: moviesViewModel.searchResults, id: \.id) { movie in
-                ForEach(allMoviewViewModel.allMovies , id: \.id) { movie in
+                ForEach(searchResults.isEmpty ?  allMoviewViewModel.allMovies: searchResults, id: \.id) { movie in
                     VStack{
                         Divider()
                         NavigationLink(destination: MovieDetailsView(viewModel: MovieDetailsViewModel(movie: movie))) {
                             MovieCellView(image: movie.image, title: movie.title, releaseDate: movie.release_date)
                         }
                         //Pagination
-                        //if(moviesViewModel.searchResults.isEmpty && self.moviesViewModel.allMovies.last?.id == movie.id){
-                        if(self.allMoviewViewModel.allMovies.last?.id == movie.id){
+                        if(searchResults.isEmpty && self.allMoviewViewModel.allMovies.last?.id == movie.id){
                             Divider()
                             Text("Fetching more...")
                                 .onAppear(perform: {
-                                    // 10 - Burası metod olmalı sadece
-                                    self.allMoviewViewModel.getTopMovies(pageNum: self.allMoviewViewModel.updateMovies())
+                                    self.allMoviewViewModel.getTopMovies(pageNum: self.allMoviewViewModel.pageNum)
                                 })
                         }
                     }
@@ -41,6 +37,6 @@ struct AllMoviesView: View {
 
 struct AllMoview_Previews: PreviewProvider {
     static var previews: some View {
-        AllMoviesView()
+        AllMoviesView(searchResults: .constant([Movie(id: 1, title: "test", release_date: "test", image: "test", voteAverage: 3.2, voteCount: 1234, overview: "ads")]))
     }
 }
